PoC: Running up a WebSocket Server on port 54321 as a ButtplugClient
Players on the local network will still get autoconnected to if running
intiface on all addresses (or if the server is on localhost), but if the
server is really remote, they can use intiface-engine like so:

   intiface-engine.exe --use-bluetooth-le --websocket-client-address ws://nsfw-minecraft-server:54321/myminecraftname

If the username specified matches a player (and that player doesn't already
have a buttplug connection), this will create a new ButtplugClient for the
user (despite origianting from a WebsocketServer).

Unfortunatly the buttplug4j packages aren't as public as I though they would be
so building is non-trivial without my credentials...

The majorty of this was hacked together this evening, so the code is far from clean.